refactor(test-kernel): merge ToolUseWaitNode satellite suite and hoist duplicated fixtures
Fold ToolUseWaitNodeFollowUpTranscriptLog.vitest.ts into ToolUseWaitNode.vitest.ts
(its module's main suite), rebuilding its services on the existing
createWaitNodeServices builder; the #7508-pattern regression cases move verbatim.

Hoist the byte-identical projectTaskGroupsFromStreamLog fold (x3) into
support/transcriptRowFixtures and the byte-identical physicistCatalog builder
plus its AgentCatalog alias (x2) into a new support/agentCatalogFixtures shared
by the extension and desktop agent-settings suites.
