+++ src/test-kernel/cli/WorkflowRunDetails.vitest.ts
@@ -23,25 +23,14 @@ import {
   type OutputFileInfo,
   STREAM_LOG_ENTRY_TYPES,
   STREAM_PHASE,
-  type StreamLogEntry,
   type StreamTabId,
   type TaskGroup,
 } from '@shared/schemas';
-import { upsertTaskGroupFromStreamLog } from '@shared/streams/taskGroupProjection';
 import { loadInk } from '@test/support/inkTestHarness.ts';
-import { textRowFixture } from '@test/support/transcriptRowFixtures';
-
-/** Test-local full replay through the production reducer (the resync path). */
-function projectTaskGroupsFromStreamLog(
-  entries: Iterable<StreamLogEntry>,
-): TaskGroup[] {
-  const taskGroups: TaskGroup[] = [];
-  const taskGroupIndex = new Map<string, number>();
-  for (const entry of entries) {
-    upsertTaskGroupFromStreamLog(taskGroups, taskGroupIndex, entry);
-  }
-  return taskGroups;
-}
+import {
+  projectTaskGroupsFromStreamLog,
+  textRowFixture,
+} from '@test/support/transcriptRowFixtures';
 
 const STREAM_ID = 'workflow#details' as StreamTabId;
